+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.6.7
 + Items with enchantability 0 can't be inserted in som enchanting machines anymore
++ Fixed some casting issues
 
 # 1.6.6
 + Made fake players not mounting entities