+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Items with enchantability 0 can't be inserted in som enchanting machines anymore
 + Fixed some casting issues
 + Fluid Pump now stop checking its area when the area is finished once
++ Fixed Book entries missing if a machines was disabled
 
 # 1.6.6
 + Made fake players not mounting entities